Added Strategy design pattern
This pattern will be used in the song model to correctly print on screen the release date based on the required precision.
Note: the format by month needs further testing because it requires the API 24 and I am not sure if "YYYY" will print the year as expected. It should, however.